+++ tests/files/src/0409_or_operator.php
@@ -34,5 +34,6 @@ function example409Reassign($scalar) {
 function exampleReassign($scalar) {
     if (is_string($scalar) || ($scalar = false)) {
         echo intdiv($scalar, 2);  // should warn because the condition ensures $scalar becomes a string. The inside of the loop won't be executed with `false`.
+        // but fails because of the contradiction
     }
 }